# src/main/resources/swagger.yaml
openapi: 3.0.2
info:
  title: HearApp
  version: v1
...
security:
  - bearerAuth: []
paths:
...
  /users/{spaceId}/can-be-invited:
    get:
      summary: get users can be invited to this space
      operationId: getUserCanBeInvitedToThisSpace
      parameters:
        - $ref: '#/components/parameters/spaceId'
      responses:
        '200':
          $ref: '#/components/responses/ListUsersResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
      tags:
        - users
  /users/{spaceId}/can-be-deleted:
    get:
      summary: get users can be deleted from this space
      operationId: getUserCanBeDeletedFromThisSpace
      parameters:
        - $ref: '#/components/parameters/spaceId'
      responses:
        '200':
          $ref: '#/components/responses/ListUsersResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
      tags:
        - users
  /spaces/invitations:
    post:
      summary: send an invitation to join in a space
      operationId: sendInvitationToMySpace
      requestBody:
        $ref: '#/components/requestBodies/UserSpaceRoleCreateRequestBody'
      responses:
        '201':
          $ref: '#/components/responses/SpaceInvitationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
      tags:
        - spaces
  /spaces/user-space-role/{spaceId}:
    delete:
      summary: unlink own user from an space
      operationId: unlinkOwnUserFromSpace
      parameters:
        - $ref: '#/components/parameters/spaceId'
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
      tags:
        - spaces
  /spaces/user-space-role:
    post:
      summary: create a new user space role
      operationId: createUserSpaceRole
      requestBody:
        $ref: '#/components/requestBodies/UserSpaceRoleCreateRequestBody'
      responses:
        '201':
          $ref: '#/components/responses/UserSpaceRoleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
      tags:
        - spaces
    delete:
      summary: delete an user space role
      operationId: deleteUserSpaceRole
      requestBody:
        $ref: '#/components/requestBodies/UserSpaceRoleCreateRequestBody'
      responses:
        '200':
          $ref: '#/components/responses/UserSpaceRoleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
      tags:
        - spaces
  /spaces/{spaceId}/voice-notes:
    get:
      summary: get a space by id fill of voice notes
      operationId: getSpaceVoiceNotesById
      parameters:
        - $ref: '#/components/parameters/spaceId'
        - $ref: '#/components/parameters/voiceNoteNameQueryParam'
      responses:
        '200':
          $ref: '#/components/responses/SpaceListVoiceNotesResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
      tags:
        - spaces
  /spaces:
    get:
      summary: get all spaces
      operationId: getSpacesByUser
      responses:
        '200':
          $ref: '#/components/responses/SpaceListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
      tags:
        - spaces
    post:
      summary: create a new space
      operationId: createSpace
      requestBody:
        $ref: '#/components/requestBodies/SpaceCreateRequestBody'
      responses:
        '201':
          $ref: '#/components/responses/SpaceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
      tags:
        - spaces
  /spaces/{spaceId}:
    get:
      summary: get a space by id
      operationId: getSpaceById
      parameters:
        - $ref: '#/components/parameters/spaceId'
      responses:
        '200':
          $ref: '#/components/responses/SpaceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
      tags:
        - spaces
    delete:
      summary: delete a space by id
      description: delete a space by id. It need a space without voices. It make a delete about UserSpaceRole to clean the database.
      operationId: deleteSpaceById
      parameters:
        - $ref: '#/components/parameters/spaceId'
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
      tags:
        - spaces
...
  /users:
    get:
      summary: getOwnUser
      operationId: getOwnUser
      responses:
        '200':
          $ref: '#/components/responses/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /users/test:
    get:
      summary: endpoint publico test
      operationId: getTest
      responses:
        '200':
          $ref: '#/components/responses/ReponseTest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/UnexpectedError'
...
components:
  requestBodies:
...
    UserSpaceRoleCreateRequestBody:
      description: Registrar un user
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UserSpaceRoleRequestDTO'
...
    SpaceCreateRequestBody:
      description: Create a new space
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SpaceCreateDTO'
  parameters:
...
    spaceId:
      name: spaceId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/genericId'
...
    voiceNoteNameQueryParam:
      name: voiceNoteNameQueryParam
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/voiceNoteNameQueryParam'
...
  responses:
    UserResponse:
      description: OK.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UserDTO'
    SpaceListVoiceNotesResponse:
      description: OK.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SpaceListVoiceNotesDTO'
    SpaceListResponse:
      description: OK.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SpaceListDTO'
...
    SpaceInvitationResponse:
      description: OK.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SpaceInvitationDTO'
    UserSpaceRoleResponse:
      description: OK.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UserSpaceRoleDTO'
    SpaceResponse:
      description: OK.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SpaceDTO'
...
    ReponseTest:
      description: OK.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TestDTO'
    ListUsersResponse:
      description: OK.
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/UserDTO'
...
    NoContent:
      description: No Content
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Problem'
          example:
            '400':
              value:
                type: 'https://api.example.com/problems/validation-error'
                status: 400
                title: Validation Error
                detail: >-
                  One or more fields are invalid. See details for more information.
                instance: 'https://api.example.com/problems/validation-error/12345'
    Unauthorized:
      description: Bad Request
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Problem'
          example:
            '400':
              value:
                type: 'https://api.example.com/problems/validation-error'
                status: 401
                title: Validation Error
                detail: >-
                  One or more fields are invalid. See details for more information.
                instance: 'https://api.example.com/problems/validation-error/12345'
    Forbidden:
      description: Bad Request
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Problem'
          example:
            '400':
              value:
                type: 'https://api.example.com/problems/validation-error'
                status: 403
                title: Validation Error
                detail: >-
                  One or more fields are invalid. See details for more information.
                instance: 'https://api.example.com/problems/validation-error/12345'
    NotFound:
      description: Bad Request
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Problem'
          example:
            '400':
              value:
                type: 'https://api.example.com/problems/validation-error'
                status: 404
                title: Validation Error
                detail: >-
                  One or more fields are invalid. See details for more information.
                instance: 'https://api.example.com/problems/validation-error/12345'
    UnexpectedError:
      description: Bad Request
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Problem'
          example:
            '400':
              value:
                type: 'https://api.example.com/problems/validation-error'
                status: 404
                title: Validation Error
                detail: >-
                  One or more fields are invalid. See details for more information.
                instance: 'https://api.example.com/problems/validation-error/12345'
  schemas:
...
    SpaceInvitationDTO:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/genericId'
        spaceId:
          $ref: '#/components/schemas/genericId'
        senderUserId:
          $ref: '#/components/schemas/genericId'
        receiverUserId:
          $ref: '#/components/schemas/genericId'
        role:
          $ref: '#/components/schemas/RoleUserSpace'
        invitationStatus:
          $ref: '#/components/schemas/InvitationStatus'
        spaceName:
          type: string
          example: 'Universe'
        senderName:
          type: string
          example: 'Mark'
        senderLastname:
          type: string
          example: 'Suckerberg'
    UserSpaceRoleDTO:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/genericId'
        userId:
          $ref: '#/components/schemas/genericId'
        spaceId:
          $ref: '#/components/schemas/genericId'
        role:
          $ref: '#/components/schemas/RoleUserSpace'
    VoiceNoteDTO:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/genericId'
        nombre:
          type: string
          example: 'Francisco'
        description:
          type: string
          example: 'Esto es una prueba de descripcion'
        duration:
          type: integer
          example: 45
        spaceName:
          type: string
          example: 'universe'
        tags:
          type: array
          items:
            type: string
            example: 'trabajo'
        codeSpaceColor:
          type: string
          example: 'AZ'
        uploadedById:
          $ref: '#/components/schemas/genericId'
        spaceId:
          $ref: '#/components/schemas/genericId'
...
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 12345
        email:
          type: string
          example: user@user.com
        firstname:
          type: string
          example: Luis
        lastname:
          type: string
          example: Perez
        isEmailActive:
          type: boolean
          example: false
        role:
          type: string
          example: USER
        tier:
          type: string
          example: FREE
...
    TestDTO:
      type: object
      properties:
        version:
          type: string
          example: 0.0.2
...
    SpaceCreateDTO:
      type: object
      properties:
        name:
          type: string
          example: my-space
        description:
          type: string
          example: my description
...
    SpaceListVoiceNotesDTO:
      type: object
      properties:
        space:
          $ref: '#/components/schemas/SpaceDTO'
        voice-notes:
          type: array
          items:
            $ref: '#/components/schemas/VoiceNoteDTO'
    SpaceListDTO:
      type: object
      properties:
        admin:
          type: array
          items:
            $ref: '#/components/schemas/SpaceDTO'
        member:
          type: array
          items:
            $ref: '#/components/schemas/SpaceDTO'
    SpaceDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: universidad
        description:
          type: string
          example: description
        colorCode:
          type: string
          example: 'AZ'
        totalVoiceNotes:
          type: integer
          example: 7
        tags:
          type: array
          items:
            type: string
            example: 'trabajo'
...
    UserSpaceRoleRequestDTO:
      type: object
      properties:
        userId:
          $ref: '#/components/schemas/genericId'
        spaceId:
          $ref: '#/components/schemas/genericId'
        role:
          $ref: '#/components/schemas/RoleUserSpace'
      required:
        - userId
        - spaceId
...
    RoleUserSpace:
      type: string
      enum:
        - ADMIN
        - MEMBER
      description: Role de los users de la app
    InvitationStatus:
      type: string
      enum:
        - PENDING
        - ACCEPTED
        - REJECTED
      description: status de las invitaciones al space
...
    Problem:
      type: object
      properties:
        error:
          type: string
    genericId:
      title: example
      description: example
      type: integer
      format: int64
      minimum: 1
      maximum: 999999999999
      example: 1
...
    voiceNoteNameQueryParam:
      type: string
      example: "audio01"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT